refactor(yomu): brief.rs の cap/topo を子モジュールへ分離 (RC-009)
516 行の brief.rs から予算キャップと topo 順序付けを pub(super) 子モジュールへ verbatim 抽出し、brief.rs を 344 行に縮小。

- src/brief/cap.rs (110): under_cap / select_drops / apply_cap
- src/brief/topo.rs (82): build_dep_graph / append_cycle_tail / kahn_positions / topo_sort
- apply_cap / topo_sort は元 pub だがワークスペース内外で未使用のため pub(super) に最小化 (cargo check --workspace で確認)
- T-708: topo_sort のファンイン (multi-dependency) 順序付けを単体テストで明示
- file-size gate allowlist から src/brief.rs を除去 (#255 自浄ラチェット)

Refs #137
